requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.18
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
email-validator==2.3.0
greenlet==3.4.0
python-dotenv==1.0.1
